The error reporting in CommandProcessor is improved. When creation of a customizer or command execution throws some exception, the user is properly notified about it. Two null pointer exceptions are fixed in ExecuteCommand and UserCommandSupport. This is a fix of issue #28508.

+++ vcscore/src/org/netbeans/modules/vcscore/cmdline/ExecuteCommand.java
@@ -354,7 +354,7 @@ public class ExecuteCommand extends Object implements VcsCommandExecutor {
     public String preprocessCommand(VcsCommand vc, Hashtable vars, String exec) {
         this.preferredExec = exec;
         fileSystem.getVarValueAdjustment().adjustVarValues(vars);
-        if (exec.indexOf(Variables.TEMPORARY_FILE) >= 0) {
+        if (exec != null && exec.indexOf(Variables.TEMPORARY_FILE) >= 0) {
             try {
                 File tempFile = File.createTempFile("VCS", "tmp");
                 tempFile.deleteOnExit();
